mod: Validator directive removed
