+++ src/lib/timepicker/public-api.ts
@@ -1,4 +1,3 @@
 export * from './timepicker.module';
-export * from './timepicker-validator';
-export * from './timepicker.helpers';
+export * from './timepicker.constants';
 export * from './timepicker';
