+++ src/lib/timepicker/timepicker.ts
@@ -16,10 +16,17 @@ import {
     ControlValueAccessor,
     FormControl,
     FormGroupDirective,
+    NG_VALIDATORS,
     NgControl,
-    NgForm
+    NgForm,
+    ValidationErrors,
+    ValidatorFn,
+    Validators
 } from '@angular/forms';
-import { noop, Subject } from 'rxjs';
+import {
+    noop,
+    Subject
+} from 'rxjs';
 
 import { coerceBooleanProperty } from '@ptsecurity/cdk/coercion';
 import {
@@ -30,25 +37,19 @@ import {
 import { McFormFieldControl } from '@ptsecurity/mosaic/form-field';
 
 import {
-    getDateFromTimeString,
-    getTimeStringFromDate,
-    decrementTime,
-    incrementTime,
+    ARROW_DOWN_KEYCODE,
+    ARROW_LEFT_KEYCODE,
+    ARROW_RIGHT_KEYCODE,
+    ARROW_UP_KEYCODE,
     DEFAULT_TIME_FORMAT,
-    getTimeEditState,
     TimeFormats,
-    getCursorPositionOfNextTimePartStart,
-    getCursorPositionOfPrevTimePartStart,
-    getParsedTimeParts
-} from './timepicker.helpers';
+    TimeParts
+} from './timepicker.constants';
 
 
-const ARROW_UP_KEYCODE: string = 'ArrowUp';
-const ARROW_DOWN_KEYCODE: string = 'ArrowDown';
-const ARROW_LEFT_KEYCODE: string = 'ArrowLeft';
-const ARROW_RIGHT_KEYCODE: string = 'ArrowRight';
-
-let uniqueComponentIdSuffix: number = 0;
+let _uniqueComponentIdSuffix: number = 0;
+let _validatorOnChange: () => void = noop;
+let _validator: ValidatorFn | null = () => null;
 
 export const MC_INPUT_VALUE_ACCESSOR =
     new InjectionToken<{ value: any }>('MC_INPUT_VALUE_ACCESSOR');
@@ -75,6 +76,8 @@ export const McTimepickerMixinBase = mixinErrorState(McTimepickerBase);
         '[disabled]': 'disabled',
         '[required]': 'required',
         '[attr.time-format]': 'timeFormat',
+        '[attr.min-time]': 'minTime',
+        '[attr.max-time]': 'maxTime',
         '[attr.value]': 'value',
         '(blur)': 'onBlur()',
         '(focus)': 'focusChanged(true)',
@@ -82,6 +85,18 @@ export const McTimepickerMixinBase = mixinErrorState(McTimepickerBase);
         '(keydown)': 'onKeyDown($event)'
     },
     providers: [
+        {
+            provide: NG_VALIDATORS,
+            useValue: {
+                validate(c) {
+                    return _validator ? _validator(c) : null;
+                },
+                registerOnValidatorChange(fn: () => void): void {
+                    _validatorOnChange = fn;
+                }
+            },
+            multi: true
+        },
         {
             provide: McFormFieldControl,
             useExisting: forwardRef(() => McTimepicker)
@@ -101,6 +116,18 @@ export class McTimepicker extends McTimepickerMixinBase
         'hh:mm:ss': '  :  :  ',
         'hh:mm': '  :  '
     };
+
+    static readonly SECONDS_PER_MINUTE: number = 59;
+    static readonly MINUTES_PER_HOUR: number = 59;
+    static readonly HOURS_PER_DAY: number = 23;
+
+    static readonly hoursMinutesSecondsRegExp =
+        new RegExp(/^([0-9]|0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]|[0-9]):([0-5][0-9]|[0-9])?$/);
+    static readonly hoursAndMinutesRegExp =
+        /^([0-9]|0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]|[0-9])?$/;
+    static readonly hoursOnlyRegExp =
+        /^([0-9]|0[0-9]|1[0-9]|2[0-3]):?$/;
+
 
     /** An object used to control when error messages are shown. */
     @Input() errorStateMatcher: ErrorStateMatcher;
@@ -187,11 +214,32 @@ export class McTimepicker extends McTimepickerMixinBase
             .map((timeFormatKey) => TimeFormats[timeFormatKey])
             .indexOf(formatValue) > -1 ? formatValue : DEFAULT_TIME_FORMAT;
 
+        _validatorOnChange();
         this.placeholder = McTimepicker.TIMEFORMAT_PLACEHOLDERS[this._timeFormat.toLowerCase()];
     }
 
+    @Input('min-time')
+    get minTime(): string | null { return this._minTime; }
+
+    set minTime(minValue: string | null) {
+        this._minTime = minValue !== null && this._getDateFromTimeString(minValue) !== undefined ?
+            minValue :
+            null;
+        _validatorOnChange();
+    }
+
+    @Input('max-time')
+    get maxTime(): string | null { return this._maxTime; }
+
+    set maxTime(maxValue: string | null) {
+        this._maxTime = maxValue !== null && this._getDateFromTimeString(maxValue) !== undefined ?
+            maxValue :
+            null;
+        _validatorOnChange();
+    }
+
     private _id: string;
-    private readonly _uid = `mc-timepicker-${uniqueComponentIdSuffix++}`;
+    private readonly _uid = `mc-timepicker-${_uniqueComponentIdSuffix++}`;
     private _disabled: boolean;
     private _required: boolean;
     private _previousNativeValue: any;
@@ -199,6 +247,8 @@ export class McTimepicker extends McTimepickerMixinBase
     private _onChange: (value: any) => void;
     private _onTouched: () => void;
     private _timeFormat: TimeFormats;
+    private _minTime: string | null = null;
+    private _maxTime: string | null = null;
 
     constructor(private readonly _elementRef: ElementRef,
                 @Optional() @Self() public ngControl: NgControl,
@@ -221,6 +271,12 @@ export class McTimepicker extends McTimepickerMixinBase
 
         // Instead of NG_VALUE_ACCESSOR (https://github.com/angular/material2/issues/8158#issuecomment-344618103)
         if (this.ngControl) { this.ngControl.valueAccessor = this; }
+
+        _validator = Validators.compose([
+            () => this._parseValidator(),
+            () => this._minTimeValidator(),
+            () => this._maxTimeValidator()
+        ]);
     }
 
     ngOnChanges(): void {
@@ -271,7 +327,7 @@ export class McTimepicker extends McTimepickerMixinBase
             hoursOnly,
             hoursAndMinutes,
             hoursAndMinutesAndSeconds
-        } = getParsedTimeParts(this._elementRef.nativeElement.value);
+        } = this._getParsedTimeParts(this._elementRef.nativeElement.value);
 
         // tslint:disable no-magic-numbers
         if (hoursOnly &&
@@ -323,7 +379,7 @@ export class McTimepicker extends McTimepickerMixinBase
         if (value !== null) {
             this._renderer.setProperty(this._elementRef.nativeElement,
                 'value',
-                getTimeStringFromDate(value, this.timeFormat)
+                this._getTimeStringFromDate(value, this.timeFormat)
             );
         }
         this._onChange(value || null);
@@ -373,13 +429,13 @@ export class McTimepicker extends McTimepickerMixinBase
     } = {}): void {
         const { changedTime, doTimestringReformat = true } = applyParams;
 
-        const timeToApply = changedTime || getDateFromTimeString(this._elementRef.nativeElement.value);
+        const timeToApply = changedTime || this._getDateFromTimeString(this._elementRef.nativeElement.value);
 
         if (doTimestringReformat && timeToApply !== undefined) {
             this._renderer.setProperty(
                 this._elementRef.nativeElement,
                 'value',
-                getTimeStringFromDate(timeToApply, this.timeFormat));
+                this._getTimeStringFromDate(timeToApply, this.timeFormat));
         }
 
         (<FormControl> this.ngControl.control).updateValueAndValidity();
@@ -391,30 +447,30 @@ export class McTimepicker extends McTimepickerMixinBase
     private _upDownTimeByArrowKeys(event: KeyboardEvent): void {
         event.preventDefault();
 
-        let changedTime: Date | undefined = getDateFromTimeString(this._elementRef.nativeElement.value);
+        let changedTime: Date | undefined = this._getDateFromTimeString(this._elementRef.nativeElement.value);
         if (changedTime !== undefined) {
             const cursorPos = this._elementRef.nativeElement.selectionStart;
 
-            const modifiedTimePart = getTimeEditState(cursorPos, this._elementRef.nativeElement.value)
+            const modifiedTimePart = this._getTimeEditState(cursorPos, this._elementRef.nativeElement.value)
                 .modifiedTimePart;
             const keyCode: string = this._getKeyCode(event);
-            if (keyCode === ARROW_UP_KEYCODE) { changedTime = incrementTime(changedTime, modifiedTimePart); }
-            if (keyCode === ARROW_DOWN_KEYCODE) { changedTime = decrementTime(changedTime, modifiedTimePart); }
+            if (keyCode === ARROW_UP_KEYCODE) { changedTime = this._incrementTime(changedTime, modifiedTimePart); }
+            if (keyCode === ARROW_DOWN_KEYCODE) { changedTime = this._decrementTime(changedTime, modifiedTimePart); }
             this._applyInputChanges({ changedTime });
             this._createSelectionOfTimeComponentInInput(cursorPos);
         }
     }
 
     private _switchSelectionBetweenTimeparts(event: KeyboardEvent): void {
-        const changedTime: Date | undefined = getDateFromTimeString(this._elementRef.nativeElement.value);
+        const changedTime: Date | undefined = this._getDateFromTimeString(this._elementRef.nativeElement.value);
         const keyCode: string = this._getKeyCode(event);
 
         if (changedTime !== undefined) {
             let cursorPos: number = this._elementRef.nativeElement.selectionStart;
             if (keyCode === ARROW_LEFT_KEYCODE) {
-                cursorPos = getCursorPositionOfPrevTimePartStart(cursorPos, this._elementRef.nativeElement.value);
+                cursorPos = this._getCursorPositionOfPrevTimePartStart(cursorPos, this._elementRef.nativeElement.value);
             } else if (keyCode === ARROW_RIGHT_KEYCODE) {
-                cursorPos = getCursorPositionOfNextTimePartStart(cursorPos, this._elementRef.nativeElement.value);
+                cursorPos = this._getCursorPositionOfNextTimePartStart(cursorPos, this._elementRef.nativeElement.value);
             }
             this._createSelectionOfTimeComponentInInput(cursorPos);
         }
@@ -429,9 +485,248 @@ export class McTimepicker extends McTimepickerMixinBase
 
     private _createSelectionOfTimeComponentInInput(cursorPos: number): void {
         setTimeout(() => {
-            const newEditParams = getTimeEditState(cursorPos, this._elementRef.nativeElement.value);
+            const newEditParams = this._getTimeEditState(cursorPos, this._elementRef.nativeElement.value);
             this._elementRef.nativeElement.selectionStart = newEditParams.cursorStartPosition;
             this._elementRef.nativeElement.selectionEnd = newEditParams.cursorEndPosition;
         });
+    }
+
+    private _incrementTime(dateVal: Date,
+                           whatToIncrement: TimeParts = TimeParts.seconds): Date {
+        let { hours, minutes, seconds } = this._getTimeDigitsFromDate(dateVal);
+
+        switch (whatToIncrement) {
+            case TimeParts.hours:
+                hours++;
+                break;
+            case TimeParts.minutes:
+                minutes++;
+                break;
+            case TimeParts.seconds:
+                seconds++;
+                break;
+            default:
+        }
+
+        if (seconds > McTimepicker.SECONDS_PER_MINUTE) { seconds = 0; }
+
+        if (minutes > McTimepicker.MINUTES_PER_HOUR) { minutes = 0; }
+
+        if (hours > McTimepicker.HOURS_PER_DAY) { hours = 0; }
+
+        return <Date> this._getDateFromTimeDigits(hours, minutes, seconds);
+    }
+
+    /**
+     * @description Decrement part of time
+     */
+    private _decrementTime(dateVal: Date,
+                           whatToDecrement: TimeParts = TimeParts.seconds): Date {
+        let { hours, minutes, seconds } = this._getTimeDigitsFromDate(dateVal);
+
+        switch (whatToDecrement) {
+            case TimeParts.hours:
+                hours--;
+                break;
+            case TimeParts.minutes:
+                minutes--;
+                break;
+            case TimeParts.seconds:
+                seconds--;
+                break;
+            default:
+        }
+
+        if (seconds < 0) { seconds = McTimepicker.SECONDS_PER_MINUTE; }
+
+        if (minutes < 0) { minutes = McTimepicker.MINUTES_PER_HOUR; }
+
+        if (hours < 0) { hours = McTimepicker.HOURS_PER_DAY; }
+
+        return <Date> this._getDateFromTimeDigits(hours, minutes, seconds);
+    }
+
+    private _getCursorPositionOfPrevTimePartStart(cursorPos: number, timeString: string): number {
+        return cursorPos === 0 ? timeString.length : cursorPos - 1;
+    }
+
+    private _getCursorPositionOfNextTimePartStart(cursorPos: number,
+                                                  timeString: string,
+                                                  timeDevider: string = ':'): number {
+        const nextDividerPos: number = timeString.indexOf(timeDevider, cursorPos);
+
+        return nextDividerPos !== undefined ? nextDividerPos + 1 : 0;
+    }
+
+    /**
+     * @description Get params for arrow-keys (up/down) time valie edit.
+     * @param cursorPosition Current cursor position in timeString
+     * @param timeString Time string to parse
+     */
+    private _getTimeEditState(cursorPosition: number, timeString: string): {
+        modifiedTimePart: TimeParts;
+        cursorStartPosition: number;
+        cursorEndPosition: number;
+    } {
+        let modifiedTimePart: TimeParts;
+        let cursorStartPosition: number;
+        let cursorEndPosition: number;
+        const hoursIndex = 0;
+        const minutesIndex = timeString.indexOf(':', hoursIndex + 1);
+        const secondsIndex = minutesIndex !== -1 ? timeString.indexOf(':', minutesIndex + 1) : -1;
+
+        if (secondsIndex !== -1 && cursorPosition > secondsIndex) {
+            modifiedTimePart = TimeParts.seconds;
+            cursorStartPosition = secondsIndex + 1;
+            cursorEndPosition = timeString.length;
+        } else if (minutesIndex !== -1 && cursorPosition > minutesIndex) {
+            modifiedTimePart = TimeParts.minutes;
+            cursorStartPosition = minutesIndex + 1;
+            cursorEndPosition = secondsIndex > -1 ? secondsIndex : timeString.length;
+        } else {
+            modifiedTimePart = TimeParts.hours;
+            cursorStartPosition = hoursIndex;
+            cursorEndPosition = minutesIndex !== -1 ? minutesIndex : timeString.length;
+        }
+
+        return {
+            modifiedTimePart,
+            cursorStartPosition,
+            cursorEndPosition
+        };
+    }
+
+    /**
+     * @description Create time string for displaying inside input element of UI
+     */
+    private _getTimeStringFromDate(tempVal: Date,
+                                   timeFormat: TimeFormats = DEFAULT_TIME_FORMAT): string {
+        const hours: string = this._getNumberWithLeadingZero(tempVal.getHours());
+        const minutes: string = this._getNumberWithLeadingZero(tempVal.getMinutes());
+        const seconds: string = this._getNumberWithLeadingZero(tempVal.getSeconds());
+
+        const formattedTimeGenerators = {
+            [TimeFormats.HHmm]: () => `${hours}:${minutes}`,
+            [TimeFormats.HHmmss]: () => `${hours}:${minutes}:${seconds}`
+        };
+
+        return formattedTimeGenerators[timeFormat]();
+    }
+
+    private _getParsedTimeParts(timeString: string): {
+        hoursOnly: any;
+        hoursAndMinutes: any;
+        hoursAndMinutesAndSeconds: any;
+    } {
+        const hoursAndMinutesAndSeconds = timeString.match(McTimepicker.hoursMinutesSecondsRegExp);
+        const hoursAndMinutes = timeString.match(McTimepicker.hoursAndMinutesRegExp);
+        const hoursOnly = timeString.match(McTimepicker.hoursOnlyRegExp);
+
+        return {
+            hoursOnly,
+            hoursAndMinutes,
+            hoursAndMinutesAndSeconds
+        };
+    }
+
+    /**
+     * @description Create Date object from separate parts of time
+     */
+    private _getDateFromTimeDigits(hours: number, minutes: number, seconds: number = 0): Date | undefined {
+        return this._getDateFromTimeString(`${hours}:${minutes}:${seconds}`);
+    }
+
+    private _getDateFromTimeString(timeString: string | undefined): Date | undefined {
+        // TODO Use moment-js
+        if (timeString === undefined) { return; }
+
+        const {
+            hoursOnly,
+            hoursAndMinutes,
+            hoursAndMinutesAndSeconds
+        } = this._getParsedTimeParts(timeString);
+
+        if (timeString.trim().length === 0 ||
+            hoursOnly === null && hoursAndMinutes === null && hoursAndMinutesAndSeconds === null) {
+            return;
+        }
+
+        // tslint:disable no-magic-numbers
+        let hours: number = 0;
+        let minutes: number = 0;
+        let seconds: number = 0;
+
+        if (hoursOnly) {
+            hours = Number(hoursOnly[1]);
+        } else if (hoursAndMinutes) {
+            hours = Number(hoursAndMinutes[1]);
+            minutes = Number(hoursAndMinutes[2]);
+        } else if (hoursAndMinutesAndSeconds) {
+            hours = Number(hoursAndMinutesAndSeconds[1]);
+            minutes = Number(hoursAndMinutesAndSeconds[2]);
+            seconds = Number(hoursAndMinutesAndSeconds[3]);
+        }
+        // const timestamp: number = Date.parse(fullDateString);
+        const resultDate: Date = new Date(1970, 0, 1, hours, minutes, seconds);
+        // tslint:enable no-magic-numbers
+
+        return isNaN(resultDate.getTime()) ? undefined : resultDate;
+    }
+
+    private _getNumberWithLeadingZero(digit: number): string {
+        const MAX_DIGIT_WITH_LEADING_ZERO: number = 9;
+
+        return digit > MAX_DIGIT_WITH_LEADING_ZERO ? `${digit}` : `0${digit}`;
+    }
+
+    private _getTimeDigitsFromDate(dateVal: Date): { hours: number; minutes: number; seconds: number } {
+        return {
+            hours: dateVal.getHours(),
+            minutes: dateVal.getMinutes(),
+            seconds: dateVal.getSeconds()
+        };
+    }
+
+    private _parseValidator(): ValidationErrors | null {
+        const dateRepresentationOfCurrentInputValue: Date | undefined =
+            this._getDateFromTimeString(this._elementRef.nativeElement.value);
+
+        return dateRepresentationOfCurrentInputValue === undefined ?
+            { mcTimepickerParse: { text: this._elementRef.nativeElement.value } } :
+            null;
+    }
+
+    private _minTimeValidator(): ValidationErrors | null {
+        const dateRepresentationOfCurrentInputValue: Date | undefined =
+            this._getDateFromTimeString(this._elementRef.nativeElement.value);
+
+        return (dateRepresentationOfCurrentInputValue !== undefined && this.minTime !== null &&
+            this._isTimeLowerThenMin(dateRepresentationOfCurrentInputValue, this.minTime)) ?
+            { mcTimepickerLowerThenMintime: { text: this._elementRef.nativeElement.value } } :
+            null;
+    }
+
+    private _maxTimeValidator(): ValidationErrors | null {
+        const dateRepresentationOfCurrentInputValue: Date | undefined =
+            this._getDateFromTimeString(this._elementRef.nativeElement.value);
+
+        return (dateRepresentationOfCurrentInputValue !== undefined && this.maxTime !== null &&
+            this._isTimeGreaterThenMax(dateRepresentationOfCurrentInputValue, this.maxTime)) ?
+            { mcTimepickerHigherThenMaxtime: { text: this._elementRef.nativeElement.value } } :
+            null;
+    }
+
+    private _isTimeLowerThenMin(timeToCompare: Date, minTime: string): boolean {
+        const minTimeForCompare: Date | undefined = this._getDateFromTimeString(minTime);
+        const timeDelta: number = timeToCompare.getTime() - (<Date> minTimeForCompare).getTime();
+
+        return timeDelta < 0;
+    }
+
+    private _isTimeGreaterThenMax(compareWithTime: Date, maxTime: string): boolean {
+        const maxTimeForCompare: Date | undefined = this._getDateFromTimeString(maxTime);
+        const timeDelta: number = compareWithTime.getTime() - (<Date> maxTimeForCompare).getTime();
+
+        return timeDelta >= 0;
     }
 }
